AoC 2025 Day 3
Part 1 was quite jank to be honest. I started by finding the largest digit in each bank, and then identified the next largest digit to the left and right of this largest digit. Merging the middle digit with both left and right while preserving order, the larger out of the two resulting integers is the largest possible integer constructed for each bank.

Part 2's solution was a general solution for finding the largest n-digit integer while preserving the order of the integers occurring in each bank of length k. For each bank, the search was done from left to right. Given that a total of (i-1) digits were found, we aim to find the largest digit from the right of the first occurrence of the previous largest digit (the (i-1)th digit) or the start of the bank (for the first search) to the (k-n+i)th integer in the bank. The reason for the (k-n+i) calculation is because for the i-th search, we know that the integer to be found cannot be any of the last (n-i) integers (else a n-digit integer cannot be formed), hence we reduce the search scope for each of the n iterations. The final integer pieced together from the digits found above is guaranteed to be the largest possible n-digit integer constructed this way, in line with the intuition of building the largest integer from a set of digits.

I think the reason why the Part 1 solution worked is because the first step was correct: find the largest integer in each bank. Once that was found, I then searched for the largest digit on the left and right of this newfound digit (for example: '...8...9...7...'). By piecing together adjacent digits and comparing them (89 vs 97), we can identify that the right integer is the larger of the two, and therefore is the largest possible 2-digit integer formed this way. What I realised while solving Part 2 was that I did not need to look at the left side if I only search for the i-th digit from the first (k-n+i) digits (or less as digits as all digits occuring before the (i-1)th digit are ignored). The only case where the left side was the largest is when the largest digit in the bank is the last digit.

All in all, it's another playing-with-numbers puzzle that I definitely enjoyed!