AoC 2025 Day 4
Today's puzzle is a simple map search puzzle. The first part requires us to find the number of accessible rolls (represented by '@') that can be removed from the map. By definition, an accessible roll is one that is surrounded by less than 4 other rolls in the map out of 8. This was achieved by iterating all positions and finding the number of rolls that satisfy this condition.

The second part goes deeper by asking for the number of rolls that can be removed from the map by iterating over the check in part 1 until no more rolls can be removed. I reused the algorithm in part 1, except that for each hit, I immediately 'remove' it from the map and increment the total count. I realised after submitting that I could also mark the locations of removed rolls with another symbol that is not '.', then counting the number of occurrences of that symbol in the resulting map. However, that would result in more list accesses  which might not be the most efficient way.