AoC 2025 Day 5
I enjoyed this puzzle involving number intervals. First part is just comparing the IDs with the limits of each interval until one that contains the ID is found.

For part 2, I initially tried to use range() and set to solve, but ran into a MemoryError. This meant I had to merge the intervals, which was easier than I initially thought. Iterating over the intervals while keeping track of the intervals that I have checked, I attempted to find any overlap of each interval with the ones that I have checked. If there is an overlap, I merge them together by taking the min and max of the two sets of limits. In the end, I ended up with non-overlapping intervals which sizes can be determined by calculating (upper-lower+1). The answer is just the sum of the sizes of the non-overlapping intervals.