AoC 2025 Day 7
Remind me next time to read the problem carefully before answering it. 😅

I initially thought part 1 asked for the number of beams formed in the end, but later realized that the required answer was the number of times the beam was split. Easy fix though, since I already have the logic to detect splits.

The idea of part 2 was quite simple as well: counting the number of unique pathways a beam can take. Initially I just replaced the set with a list, but then found that the code ran for way too long. Soon I realized that the list probably exploded in size, so I changed to a dict-based counter method that kept track of how many beams were in each position. This made the algorithm much more efficient, and it completed as soon as I executed it.

I'll probably start timing my algorithms, maybe after this year's AoC is done and I have more time on my hands.