AoC 2025 Day 8
From today onwards, I will time my algorithms and print them out in the console.

This might be the toughest problem so far for AoC 2025. It took me a few hours to get the algorithm right. This puzzle reminded me of AoC 2024 Day 12 which was about regions, as today's puzzle is about connected circuits.

Initially, I thought I have to implement some graph-searching algorithm like I did for AoC 2024 Day 12, but I managed to find an alternative solution: organizing points into sets that represent connected circuits (single-box circuits are omitted). After identifying all unique edges and sorting them by increasing edge lengths, I iterated over the sorted list and kept track of the connected circuits that formed.

For each edge, I check if each end point already belong to any of the existing circuits. If there are 2 sets that contain at least 1 of the 2 endpoints (which is only possible if one set contained 1 end point and the second set contained the other), the two sets are merged. If only 1 set contained an endpoint, then I add both endpoints to that set. Otherwise, I create a new circuit with that edge. For efficiency, I only operate on the index of the points.

Part 1 only required the shortest 1000 edges out of the 1000*999/2 = 499500 unique edges to be connected, and asked for the product of the sizes of the top 3 largest circuits. For Part 2, I kept iterating until all circuit boxes are connected in one single circuit. From there, I identified the endpoints of the last edge and got the product of their x-coordinates as the solution.

Both Parts 1 and 2 took just under 0.5s. Might not be the optimal solution, but it worked out quite well.