AoC Day 10 Part 1
For the first time this year, I am unable to complete both parts in 1 day :(

Part 1 involves a simple BFS (floodfill?) to find the minimum number of button presses to end up with the required configurations.

I attempted a similar method in Part 2 which theoretically should work. However to my dismay, the number of combinations explored at each iteration increased exponentially even after disregarding duplicate sums and sums that exceed the target joltages. I think BFS worked only in Part 1 was because that the search space was much smaller compared to Part 2 (2^n where n=number of machines compared to the product of joltages).

Currently looking at tips on Reddit and found a method using matrix rref. Might be feasible but need more work.

In the meantime, I'm gonna catch up on sleep as I have a meeting to attend tomorrow morning.