AoC 2025 Day 11 Part 1
Another day of unable to solve Part 2 by midnight :(

Part 1 is quite straightforward, just another BFS to find the number of paths from 'you' to 'out'.

I initially modified Part 1's solution for Part 2, but that took forever to run. Even with my latest attempt (not shown) with memoization, it still took over 20 minutes and ended with a pesky MemoryError. Guess I'll have to find a better way to solve this soon...