AoC 2025 Day 12
What a emotional rollercoaster ride today! I was fully prepared to brute-force all the possible present layouts, which took FOREVER just on the first config. I thought that if AoC was still 25 days long, I wouldn't be able to keep up. Only much later did I find that the inputs were all special cases (either the total areas exceed the grid area or the total number of 3x3 boxes for each present covers less than or equal to the total grid area. Just to be sure, I left my original brute-force algorithm in to see if that branch ever gets triggers, and NONE of the inputs triggered that branch :(

And to top it all off, Part 2 was free! What a dramatic end to my AoC 2025 experience!